requirements.txt
streamlit
google-generativeai
pyMuPDF
python-dotenv
llama-index
llama-index-storage-docstore-mongodb
llama-index-storage-index-store-mongodb
motor
llama_index-llms-ollama
llama_index-embeddings-huggingface
llama-index-vector-stores-chroma
openai
llama-index-graph-stores-neo4j
unstructured[pdf]
unstructured.PaddleOCR
paddlepaddle 